refactored the code and made some changes to checkins service
