+++ server/src/main/java/com/objectcomputing/checkins/services/checkins/CheckInsController.java
@@ -1,4 +1,4 @@
-package com.objectcomputing;
+package com.objectcomputing.checkins.services.checkins;
 
 import java.net.URI;
 import java.util.List;
@@ -7,7 +7,7 @@ import java.util.UUID;
 import javax.annotation.Nullable;
 import javax.validation.Valid;
 
-import com.objectcomputing.checkins.CheckIns;
+import com.objectcomputing.checkins.services.checkins.CheckIns;
 
 import io.micronaut.http.HttpResponse;
 import io.micronaut.http.annotation.Body;
